refactor(desktop): extract install command execution into install_exec
agent_discovery.rs sat at its 2167-line ratchet ceiling with no headroom
for the install supervision and log-persistence work queued behind it.
Install command execution — spawn, output drain under timeout, retry with
backoff, output truncation — is a self-contained seam and moves out with
its tests, matching the existing managed_node.rs / post_install_verification.rs
split. Command construction stays in the parent; the new module owns only
what happens once a Command exists, and exposes exactly one pub(super) fn.

Installers that write relative to the working directory failed under a
packaged desktop launch, which inherits / as its CWD (#2245, absorbed from
#3090). run_install_command now goes through prepare_install_command, which
applies default_agent_workdir() to every command it builds, so no spawn
site can bypass it.
